Turning, states, renaming, rumble, POV, manual angling, angle up, angle down
Changed the turning to max speed. Added states; they are for changing which motors will be used for the right joystick. We also renamed the motors for the arms to DynamicArm1 and DynamicArm2. We added rumbling to the controller when it changes states. The d-pad is the pov angles, added constants for those. Added manual angling command. Added angle up command. Added angle down command.